==> 3.2_feature_engineering.ipynb <==
## Import necessary libraries

# Data imports
import pandas as pd

# Prepare environment

INTERMEDIATE_SAVE_PATH = "data/big_matrix_intermediate_saves/"
CLEANED_DATA_PATH = "data/cleaned/"

# Load Data

final_data = pd.read_parquet(INTERMEDIATE_SAVE_PATH + "final_data.parquet")
user_features_cleaned = pd.read_parquet(CLEANED_DATA_PATH + "user_features_cleaned.parquet")

# Data preprocess for AI

### Final data cleanup and merging of all user features

# Copy Dataframe
final_data = final_data.copy()

# Merge all user features
final_data = final_data.merge(user_features_cleaned, on="user_id", how="left")

# Encode columns to be easier to work with
final_data['user_active_degree'] = pd.factorize(final_data['user_active_degree'])[0]
final_data['follow_user_num_range'] = pd.factorize(final_data['follow_user_num_range'])[0]
final_data['fans_user_num_range'] = pd.factorize(final_data['fans_user_num_range'])[0]
final_data['friend_user_num_range'] = pd.factorize(final_data['friend_user_num_range'])[0]
final_data['register_days_range'] = pd.factorize(final_data['register_days_range'])[0]

final_data

# Save Data

final_data.to_parquet(INTERMEDIATE_SAVE_PATH + "merged.parquet", index=True)
