==> 3.4_feature_engineering copy.ipynb <==
## Import necessary libraries

# Data imports
import numpy as np
import pandas as pd
from pathlib import Path
import utils
#from utils import compute_averages, merge_matrix_avgs_friend_list, get_friends_watch_ratio

# Prepare environment

SEED = 42
np.random.seed(SEED)

TRAIN_TEST_SPLIT = 0.8
CORR_THRESHOLD = 0.2

CLEANED_DATA_PATH = "data/cleaned/"
FEATURES_PATH = "data/features/"
Path(FEATURES_PATH).mkdir(parents=True, exist_ok=True)

# Load Data

big_matrix_final = pd.read_parquet(FEATURES_PATH + "big_matrix_final.parquet")
avg_user_feat = pd.read_parquet(FEATURES_PATH + "avg_user_feat.parquet")
avg_user_category = pd.read_parquet(FEATURES_PATH + "avg_user_category.parquet")

small_matrix_cleaned = pd.read_parquet(CLEANED_DATA_PATH + "small_matrix_cleaned.parquet")
user_features_cleaned = pd.read_parquet(CLEANED_DATA_PATH + "user_features_cleaned.parquet")
social_network_cleaned = pd.read_parquet(CLEANED_DATA_PATH + "social_network_cleaned.parquet")
item_categories_cleaned = pd.read_parquet(CLEANED_DATA_PATH + "item_categories_cleaned.parquet")
caption_category_cleaned = pd.read_parquet(CLEANED_DATA_PATH + "caption_category_cleaned.parquet")

# Check Dataframes

small_matrix_cleaned

big_matrix_final

user_features_cleaned

social_network_cleaned

item_categories_cleaned

# Data preprocess for AI

### Final data cleanup and merging of all user features

# Copy Dataframe
final_data = small_matrix_cleaned.copy()

video_feat_category_avg_per_user = utils.get_video_feat_category_avg_per_user(final_data, avg_user_feat, avg_user_category, item_categories_cleaned, caption_category_cleaned)

final_data = final_data.merge(video_feat_category_avg_per_user, on=["user_id", "video_id"], how="left")

final_data["user_feat_watch_ratio_mean"] = final_data["user_feat_watch_ratio_mean"].fillna(0)
final_data["user_category_watch_ratio_mean"] = final_data["user_category_watch_ratio_mean"].fillna(0)

final_data

# Merge user calculated data from big_matrix
user_big_matrix_final = big_matrix_final[["user_id", "user_watch_ratio_mean"]].drop_duplicates()
user_big_matrix_final = user_big_matrix_final[user_big_matrix_final["user_id"].isin(final_data["user_id"])].drop_duplicates()
final_data = final_data.merge(user_big_matrix_final, on="user_id", how="left")

final_data.isna().sum()

## Only keep meaningful features

### Get correlation between watch_ratio and every other feature

corr = final_data.corr()["watch_ratio"]
corr

### See which features are above set threshold

corr[abs(corr) > CORR_THRESHOLD]

### Only keep previously selected features as well as useful ones for later use

final_data = final_data[big_matrix_final.columns]

# Save Data for AI

final_data.to_parquet(FEATURES_PATH + "small_matrix_final.parquet", index=True)
