==> 4_ai.ipynb <==
# Import necessary libraries

# Data imports
import pandas as pd
import numpy as np

# AI imports
import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense, Dropout, BatchNormalization, Input, Lambda

from sklearn.preprocessing import RobustScaler

# Prepare environment

SEED = 42
tf.random.set_seed(SEED)

TRAIN_TEST_SPLIT = 0.8
FEATURES_PATH = "data/features/"

# Prepare Data

## Load Data

scaler = RobustScaler()

big_matrix_final = pd.read_parquet(FEATURES_PATH + "big_matrix_final.parquet")
black_list = ["user_id", "video_id", "watch_ratio"]
to_scale = [col for col in big_matrix_final.columns if col not in black_list]
big_matrix_final[to_scale] = scaler.fit_transform(big_matrix_final[to_scale])
big_matrix_final = big_matrix_final.drop(columns=["user_id", "video_id"])


small_matrix_final = pd.read_parquet(FEATURES_PATH + "small_matrix_final.parquet")
small_matrix_final[to_scale] = scaler.transform(small_matrix_final[to_scale])
small_matrix_final.to_parquet(FEATURES_PATH + "small_matrix_scaled.parquet")
small_matrix_final = small_matrix_final.drop(columns=["user_id", "video_id"])



big_matrix_final

## Split data into X_train / y_train / X_test / y_test

def split_X_y(df: pd.DataFrame):
    y = df["watch_ratio"]
    tmp = df.drop(columns=["watch_ratio"])
    X = tmp[[c for c in tmp.columns]]
    return X, y

def split_all(df: pd.DataFrame):
    train, test = big_matrix_final, small_matrix_final
    X_train, y_train = split_X_y(train)
    X_test, y_test = split_X_y(test)
    test_watch_ratio_mean = test["video_watch_ratio_mean"]
    return np.array(X_train), np.array(y_train), np.array(X_test), np.array(y_test), test_watch_ratio_mean

X_train, y_train, X_test, y_test, test_watch_ratio_mean = split_all(big_matrix_final)

# AI

## Create the model

model = Sequential([
    Input(shape=(5,)),
    
    # Block 1
    Dense(64, activation='relu'),
    BatchNormalization(),
    Dropout(0.2),

    Dense(1, activation='sigmoid'),
    # Scale output to [0, 5] to match 
    Lambda(lambda x: x * 5.0)
])

model.compile(
    optimizer='adam',
    loss='mse',
    metrics=['mae']
)

## Train the model

model.fit(
    X_train,
    y_train,
    validation_split=0.1,
    epochs=5,
    batch_size=1024,
    callbacks=[tf.keras.callbacks.EarlyStopping(patience=5, restore_best_weights=True)],
    verbose=1
)

## Predict video watch_ratio and compare it to the baseline

The baseline is the average watch_ratio of all videos.\
We can see that the MAE of my model is arounf 25% better than the baseline.

baseline = np.full_like(y_test, np.mean(y_test), dtype=np.float32)
preds = model.predict(X_test, batch_size=2048, verbose=None)
mae_baseline = np.mean(np.abs(baseline - y_test))
mae_model = np.mean(np.abs(preds.flatten() - y_test))
print(f"Baseline MAE: {mae_baseline}")
print(f"Actual MAE: {mae_model}")

# Save the model

model.save(FEATURES_PATH + "ai_model.keras")
