==> 5_recommender_system.ipynb <==
# Import necessary libraries

from tensorflow.keras.models import load_model
import numpy as np
import pandas as pd
from sklearn.metrics import ndcg_score
from tqdm import tqdm

# Prepare environment

FEATURES_PATH = "data/features/"

# Load Data

model = load_model(FEATURES_PATH + "ai_model.keras", safe_mode=False)
small_matrix_final = pd.read_parquet(FEATURES_PATH + "small_matrix_scaled.parquet")

# Check Data

small_matrix_final

# Recommender system

### Define function to get recomendations of a given user

def get_user_recommendation_data(user: int):
    preds = model.predict(small_matrix_final[small_matrix_final["user_id"] == user][[col for col in small_matrix_final.columns if col not in ["watch_ratio", "user_id", "video_id"]]], batch_size=2048, verbose=None).flatten()
    user_mat = small_matrix_final[small_matrix_final["user_id"] == user].reset_index(drop=True)
    to_recommend_index = np.flip(preds.argsort())
    to_recommend = user_mat["video_id"][to_recommend_index].values

    ground_truth_index = np.flip(user_mat["watch_ratio"].values.argsort())
    ground_truth = user_mat["video_id"][ground_truth_index].values

    return to_recommend, ground_truth

### Predict top 10 videos to recommend to some user

sample_user = 946
preds, ground_truth = get_user_recommendation_data(sample_user)
print(f"predictions: {preds[:10]}")
print(f"ground_truth: {ground_truth[:10]}")

### Calculate ndcg score for this user

ndcg_score([ground_truth], [preds], k=10)

# Evaluating metrics

### Evaluate mean NDCG metric on all users
We are using NDCG@K because it is a metrics that looks into the ranking of the predictions

ndcg_5_list = []
for user_id in tqdm(small_matrix_final["user_id"].unique()):
    user_df = small_matrix_final[small_matrix_final["user_id"] == user_id]
    recommendations, ground_truth = get_user_recommendation_data(user_id)
    
    if len(ground_truth) >= 5:
        ndcg_5_list.append(ndcg_score([ground_truth], [recommendations], k=10))

avg_ndcg_5 = np.mean(ndcg_5_list)
print(f"Average NDCG@5 over users: {avg_ndcg_5:.4f}")

### Create function to predict recommendations using the video's average watch_ratio as baseline

def get_user_recommendation_data_baseline(user: int):
    preds = np.array(small_matrix_final[small_matrix_final["user_id"] == user]["video_watch_ratio_mean"].to_list())
    user_mat = small_matrix_final[small_matrix_final["user_id"] == user].reset_index(drop=True)
    to_recommend_index = np.flip(preds.argsort())
    to_recommend = user_mat["video_id"][to_recommend_index].values

    ground_truth_index = np.flip(user_mat["watch_ratio"].values.argsort())
    ground_truth = user_mat["video_id"][ground_truth_index].values

    return to_recommend, ground_truth

### Check the NDCG score using our baseline

ndcg_5_list = []
for user_id in tqdm(small_matrix_final["user_id"].unique()):
    user_df = small_matrix_final[small_matrix_final["user_id"] == user_id]
    recommendations, ground_truth = get_user_recommendation_data_baseline(user_id)
    
    if len(ground_truth) >= 5:
        ndcg_5_list.append(ndcg_score([ground_truth], [recommendations], k=10))

avg_ndcg_5 = np.mean(ndcg_5_list)
print(f"Average NDCG@5 over users: {avg_ndcg_5:.4f}")

As we can see, our ai model is approximately on the same level as our baseline, which means that it didn't succeed in learning complex correlations between the different data.\
Unfortunately, because of my computer's limitations, I wasn't able to include all desired features into this model. Maybe further research and data analysis could have lead to a more optimized approach.

# Looking into other acceptable metrics

As we saw before, our model isn't really efficient.\
However, the ranking of every single video isn't really the most important part of a video recommender system such as Kuaishou or Tiktok.\
In fact, the important part is to be able to recommend good videos, and avoid bad ones.

Having this in mind, we can group the videos with similar watch ratios and ignore the exact ranks between videos of similar rank.\
To achieve this, we will round the predicted watch ratio, leaving us with a watch ratio belonging to this set: {0, 1, 2, 3, 4, 5}

While I know that this has limitations (like the fact that a watch ratio of 1.49 and 1.51 will be considered very differently) and that this method doesn't improve our ai model, it is still an acceptable metric.

In the following code, the videos will be sorted by rounded watch ratio and then by their video id. This last part is to be able to evaluate its ranking with our ground_truth.\
In reality, we could recommend any video within the best watch ratio predicted

def get_user_recommendation_data_rounded(user: int):
    user_mat = small_matrix_final[small_matrix_final["user_id"] == user].reset_index(drop=True)
    preds = model.predict(user_mat[[col for col in small_matrix_final.columns if col not in ["watch_ratio", "user_id", "video_id"]]], batch_size=2048, verbose=None).flatten()

    user_mat["preds_rounded"] = pd.Series(np.around(preds))
    user_mat["watch_ratio_rounded"] = pd.Series(np.around(user_mat["watch_ratio"].to_numpy()))

    to_recommend = user_mat.sort_values(["preds_rounded", "video_id"])["video_id"]
    ground_truth = user_mat.sort_values(["watch_ratio_rounded", "video_id"])["video_id"]

    return to_recommend, ground_truth

### Evaluate mean NDCG metric on all users

ndcg_5_list = []
for user_id in tqdm(small_matrix_final["user_id"].unique()):
    user_df = small_matrix_final[small_matrix_final["user_id"] == user_id]
    recommendations, ground_truth = get_user_recommendation_data_rounded(user_id)
    
    if len(ground_truth) >= 5:
        ndcg_5_list.append(ndcg_score([ground_truth], [recommendations], k=10))

avg_ndcg_5 = np.mean(ndcg_5_list)
print(f"Average NDCG@5 over users: {avg_ndcg_5:.4f}")

This improved result tells us that our ai model, while not being able to rank every single video efficiently enough, is able to somewhat tell apart good videos from bad videos.

def get_user_recommendation_data_baseline_bis(user: int):
    user_mat = small_matrix_final[small_matrix_final["user_id"] == user].reset_index(drop=True)
    preds = np.array(user_mat["video_watch_ratio_mean"].to_list())
    
    user_mat["preds_rounded"] = pd.Series(np.around(preds))
    user_mat["watch_ratio_rounded"] = pd.Series(np.around(user_mat["watch_ratio"].to_numpy()))
    
    to_recommend = user_mat.sort_values(["preds_rounded", "video_id"])["video_id"]
    ground_truth = user_mat.sort_values(["watch_ratio_rounded", "video_id"])["video_id"]

    return to_recommend, ground_truth

### Check the NDCG score using our baseline

ndcg_5_list = []
for user_id in tqdm(small_matrix_final["user_id"].unique()):
    user_df = small_matrix_final[small_matrix_final["user_id"] == user_id]
    recommendations, ground_truth = get_user_recommendation_data_baseline_bis(user_id)
    
    if len(ground_truth) >= 5:
        ndcg_5_list.append(ndcg_score([ground_truth], [recommendations], k=10))

avg_ndcg_5 = np.mean(ndcg_5_list)
print(f"Average NDCG@5 over users: {avg_ndcg_5:.4f}")

Seing as our baseline (using the video's global average watch_ratio) is still quite close to our model's metrics, we can conclude that ou ai model wasn't able to get a good grasp of a user's specific feature and is only based on the video's global popularity.
